Move ONS Components to correct folder
