+++ src/SignIn.tsx
@@ -1,9 +1,9 @@
 import React, {EventHandler, useState} from "react";
 import {Link, Redirect, useParams, useLocation} from "react-router-dom";
-import {ONSPasswordInput} from "./Components/ONSPasswordInput";
+import {ONSPasswordInput} from "./Components/ONSDesignSystem/ONSPasswordInput";
 import {isDevEnv} from "./Functions";
-import {ONSButton} from "./Components/ONSButton";
-import {ONSTextInput} from "./Components/ONSTextInput";
+import {ONSButton} from "./Components/ONSDesignSystem/ONSButton";
+import {ONSTextInput} from "./Components/ONSDesignSystem/ONSTextInput";
 
 interface Props {
     setAuthenticationToken: any
